+++ src/geo/utils.ts
@@ -2,7 +2,7 @@ import formatcoords from "formatcoords";
 import dayjs from "dayjs";
 
 export function formatDateString(value?: number, timeZone?: string) {
-  if (value === undefined) return "";
+  if (value === undefined || value === null) return "";
   if (timeZone) return dayjs(value).tz(timeZone).format();
 
   return dayjs.utc(value).format();