+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Edit visibility of multiple features at once.
+- Add button for clearing feature visibility timestamp.
 - On zoom to a unit with no location zoom to show subordinates instead.
 - Add button to the unit panel for setting unit location by clicking on the map.
 - Zoom to multiple features using z keyboard shortcut or the zoom button on the feature details panel.